backend/djangopj/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astroid==2.3.3
astunparse==1.6.3
autopep8==1.5
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cmake==3.17.2
colorama==0.4.3
defusedxml==0.6.0
distro==1.5.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
Faker==4.0.2
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
icrawler==0.6.2
idna==2.9
isort==4.3.21
joblib==0.15.1
jsonfield==3.1.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-build==0.11.1
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.11.2
